Initialize default values for flags.
This is why it was failing in the last commit. I assumed it would just 
pick up the default value of the variable (and it seems that some 
functions of pflags do work this way, but no the ones I am using 
currently). Now it initializes to the proper values.