Make bundles hashed values.
Even though they are meant to be easy to create and throw away, having a 
hashed value for bundle contents makes it easy to ensure that the 
commits that are included are all accounted for by rehashing the 
necessary values.

Add semantic version of program as a single external text file that is 
embedded into the executable.

Change the program flags to be placed into an anonymous struct so that 
they are better namedspaced per sub-command.