Run most sqlite files through `sqlfluff` formatter
The two big `tables` files are unparseable by `sqlfluff`
which is a bad sign.
These are the two files not touched by `sqlfluff`.

Sqlite is probably being very permissive in what it allows
since these scripts have not been failing when run in tests.
These need to be refactored until they are in good form
and can be linted/formatted by `sqlfluff`.