Start porting to Zig 0.15.1
This commit will not build right now. The port is not finished.

The main issue is that seems it isn't possible to make a writer out of a
union because there cannot be a dedicated field to hold the `Writer`
pointer, which is what Writer expects. I will need to refactor the
Hasher type to make it a true interface/struct instead of using a union.