Streaming is working with new Io system
I'm still not certain I am using it correctly. The tests for the add
command need to be more robust before I can be certain that this is
working correctly. I could check the results manually, but that will
eventually get tedious to inspect the sqlite DB manually for every
change.